corrections

+++ 0x12-javascript-warm_up/9-add.js
@@ -4,8 +4,8 @@ function add (a, b) {
   console.log(a + b);
 }
 
-const a = parseInt(process.argv[2]);
-const b = parseInt(process.argv[3]);
+const a = parseFloat(process.argv[2]);
+const b = parseFloat(process.argv[3]);
 if (isNaN(a) || isNaN(b)) {
   // nothing
 } else {
